refactor: moves the neo4j sessions server side to improive security
